requirements.txt
Django==3.1
django-rest-framework==0.1.0

